settings-previews to new file (caption/scripture)
